+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 * Added `fromSimple_`.
 
+* Added `injectFailures` and `keepCallStack`.
+
 ## 0.5.0.0 
 
 * The Advice type has changed to get rid of the existential type.